+++ packages/web/components/IndexingErrors/index.ts
@@ -1,0 +1,2 @@
+export { default as IndexingErrors } from "./IndexingErrors";
+export { useIndexingErrors } from "./useIndexingError";
