+++ packages/web/package.json
@@ -51,7 +51,6 @@
     "@types/react-dom": "18.2.7",
     "@types/react-timeago": "^4.1.3",
     "@types/react-virtualized": "^9.22.0",
-    "@types/react-window": "^1.8.8",
     "chai": "^5.1.1",
     "mocha": "^10.5.2",
     "tailwindcss": "3.3.3",
